Fixes #2143: Move InteractionObjectTestBuilder to a main-level testing library (#5996)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
Fix #2143 

- Move `InteractionObjectTestBuilder.kt` to
`app/src/main/java/org/oppia/android/app/player/state/testing/`.
- Update all imports using the `InteractionObjectTestBuilder.kt` file.
- Move the related `kt_android_library` code blocks from
`domain/BUILD.bazel` to `app/BUILD.bazel`.
- Update `scripts/assets/kdoc_validity_exemptions.textproto` and
`scripts/assets/test_file_exemptions.textproto` accordingly to change
file paths to `InteractionObjectTestBuilder.kt`.

<!--
- Explain what your PR does. If this PR fixes an existing bug, please
include
- "Fixes #bugnum:" in the explanation so that GitHub can auto-close the
issue
  - when this PR is merged.
  -->

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [ ] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).
