+++ app/BUILD.bazel
@@ -695,6 +695,17 @@ kt_android_library(
 #  - Only test files depend on it
 
 kt_android_library(
+    name = "interaction_object_test_builder",
+    testonly = True,
+    srcs = [
+        "src/main/java/org/oppia/android/app/player/state/testing/InteractionObjectTestBuilder.kt",
+    ],
+    deps = [
+        "//model/src/main/proto:question_java_proto_lite",
+    ],
+)
+
+kt_android_library(
     name = "test_deps",
     testonly = True,
     srcs = [
