+++ domain/BUILD.bazel
@@ -164,26 +164,14 @@ kt_android_library(
     ],
 )
 
-# TODO(#2143): Move InteractionObjectTestBuilder to a testing package outside the test folder.
-kt_android_library(
-    name = "interaction_object_test_builder",
-    testonly = True,
-    srcs = [
-        "src/test/java/org/oppia/android/domain/classify/InteractionObjectTestBuilder.kt",
-    ],
-    deps = [
-        "//model/src/main/proto:question_java_proto_lite",
-    ],
-)
-
 # keep sorted
 TEST_DEPS = [
     ":domain",
-    ":interaction_object_test_builder",
     "//:dagger",
     "//app:crashlytics",
     "//app:crashlytics_deps",
     "//app:firestore_deps",
+    "//app:interaction_object_test_builder",
     "//app/src/main/java/org/oppia/android/app/application/testing:testing_build_flavor_module",
     "//config/src/java/org/oppia/android/config:all_languages_config",
     "//data/src/main/java/org/oppia/android/data/backends/gae:network_config_prod_module",
